+++ app/src/main/java/com/adammcneilly/mvwtf/MVWTFApp.kt
@@ -1,0 +1,5 @@
+package com.adammcneilly.mvwtf
+
+import android.app.Application
+
+class MVWTFApp : Application()
