+++ app/src/main/java/com/adammcneilly/mvwtf/theme/Color.kt
@@ -1,6 +1,6 @@
 @file:Suppress("MagicNumber")
 
-package template.theme
+package com.adammcneilly.mvwtf.theme
 
 import androidx.compose.ui.graphics.Color
