+++ app/src/main/java/com/adammcneilly/mvwtf/theme/Type.kt
@@ -1,4 +1,4 @@
-package template.theme
+package com.adammcneilly.mvwtf.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
