--- app/src/main/java/template/TemplateApp.kt
@@ -1,7 +1,0 @@
-package template
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class TemplateApp : Application()
